API test setup working
