+++ .gitignore
@@ -5,6 +5,8 @@ reports/
 **/*.tsbuildinfo
 **/.env
 
+mongod-x64-pop-5.0.3 
+
 .yarn/*
 !.yarn/patches
 !.yarn/plugins
