+++ packages/server/jest.config.js
@@ -4,5 +4,5 @@ const package = require('./package.json')
 module.exports = {
   ...base,
   name: package.name,
-  displayName: package.name
+  displayName: package.name,
 }
