+++ packages/server/package.json
@@ -7,6 +7,8 @@
     "@types/node": "^16.9.2",
     "@types/url-join": "^4.0.1",
     "@types/uuid": "^8.3.1",
+    "graphql-request": "^3.7.0",
+    "mongodb-memory-server": "^8.2.0",
     "tsc-watch": "^4.5.0"
   },
   "main": "build/server.js",
